Move console_log and epoch management to its own file, logging.ts
